Rename file: redis facade to redis client facade
